requirements.txt
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
apturl==0.5.2
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
automata-lib==8.1.0
bcrypt==4.0.1
beautifulsoup4==4.10.0
beniget==0.4.1
binwalk==2.3.3
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
bs4==0.0.1
cached_method==0.1.0
caffeine==2.9.8
certifi==2023.11.17
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
comm==0.2.0
command-not-found==0.3
contourpy==1.1.0
cryptography==39.0.2
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.8.0
decorator==4.4.2
defer==1.0.6
dill==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==4.2.4
django-admin==2.0.2
django-cors-headers==4.2.0
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.14.0
dnspython==2.4.2
duplicity==0.8.21
ecdsa==0.18.0
Elixir==0.7.1
email-validator==2.1.0.post1
ewmh==0.1.6
excel-base==1.0.4
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.109.0
fasteners==0.14.1
fonttools==4.42.0
frozendict==2.3.8
fs==2.4.12
future==0.18.2
gast==0.5.2
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
ipykernel==6.27.1
ipython==8.18.1
isoweek==1.3.3
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.0
keyboard==0.13.5
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.3
nest-asyncio==1.5.8
netifaces==0.11.0
networkx==3.2
numpy==1.25.2
oauthlib==3.2.0
olefile==0.46
outcome==1.3.0.post0
packaging==23.1
pandas==1.5.1
paramiko==2.9.3
parso==0.8.3
passlib==1.7.4
pbr==5.11.1
pdfminer.six==20221105
pdfplumber==0.10.3
pexpect==4.8.0
Pillow==10.0.0
platformdirs==4.1.0
ply==3.11
prompt-toolkit==3.0.41
protobuf==3.12.4
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
PyMySQL==1.0.3
PyNaCl==1.5.0
pynvim==0.4.2
PyOpenGL==3.1.5
pyparsing==3.0.9
pypdf==3.17.4
pypdfium2==4.25.0
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyqtgraph==0.12.4
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-jose==3.3.0
python-multipart==0.0.6
python-xlib==0.29
pythran==0.10.0
pytz==2023.3
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.2
qiskit==0.44.1
qiskit-aer==0.12.2
qiskit-terra==0.25.1
reportlab==3.6.8
requests==2.25.1
rsa==4.9
rustworkx==0.13.1
scipy==1.11.2
screen==1.0.1
SecretStorage==3.3.1
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.11.0
sqlparse==0.4.4
ssh-import-id==5.11
stack-data==0.6.3
starlette==0.35.1
stevedore==5.1.0
symengine==0.9.2
sympy==1.12
systemd-python==234
Tempita==0.5.2
TimeConvert==3.0.13
tornado==6.4
traitlets==5.14.0
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.9.0
tzlocal==5.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==2.1.0
usb-creator==0.3.7
uvicorn==0.20.0
wadllib==1.3.6
wcwidth==0.2.12
webencodings==0.5.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
xlrd==0.7.1
xlwt==0.7.2
zipp==1.0.0
